Add Thawani config, update package requirements, and enhance service provider
A configuration file was added for Thawani support in package/Thawani/src/config. The package's composer.json requirement was updated to version 2.* for phpawcom/thawani-php-class. In ThawaniServiceProvider, configuration merging and publishing were enabled. Also, the README was updated with installation instructions and composer.lock was added to the .gitignore file.